+++ CHANGELOG.md
@@ -13,7 +13,7 @@ See [GitHub releases](https://github.com/mll-lab/php-utils/releases).
 
 ### Changed
 
-- Refactored `MLL\Utils\PHPStan\Rules\VariableNameIdToIDRule` to `MLL\Utils\PHPStan\Rules\NameIdToIDRule` to handle variable names, method names and class names for incorrect capitalization of "Id"
+- Refactored `MLL\Utils\PHPStan\Rules\VariableNameIdToIDRule` to `MLL\Utils\PHPStan\Rules\NameIdToIDRule` to handle variable names, parameter names, method names and class names for incorrect capitalization of "Id"
 - Add `Idt` to the list of incorrect capitalizations for "Id" of `MLL\Utils\PHPStan\Rules\NameIdToIDRule`
 
 ## v5.8.0